Update the "origin story"
I believe the proposed changes: 1) correct an English language inconsistency, 2) explain the origins of the original, 3) clarifies this library's relationship to the alternative TensorFlow library, and 4) at least hints at the benefits of using this library compared to the TensorFlow alternative. As a new user, I was confused about all of these things, and had to do my own research. I've based the proposed changes based on what I learned from TensorFlow issue #33748, so if any of it is incorrect, please let me know, or, of course, feel free to make changes.